+++ js/data.js
@@ -1,0 +1,34 @@
+/* мапы для заполнения */
+const PHOTO_NUMBER = 25;
+const MAX_COMMENT_NUMBER = 30;
+const PHOTO_DESCRIPTIONS = ['Вся красота мира в одной картинке',
+  'Моменты, которые запечатлены навсегда',
+  'Счастье в каждом кадре',
+  'Когда слова не нужны, достаточно фотографии',
+  'История, рассказанная через объектив',
+  'Остановить время в одном кадре',
+  'Фотография — это способ улыбнуться в будущем',
+  'Сегодня — самый лучший день',
+  'Я не доверяю словам. Я доверяю фотографиям',
+  'Фотографии — это свидетельство о том, что мы жили',
+  'Момент, когда небо и земля сливаются воедино',
+  'В объектив всегда видна правда — это как детектор лжи',
+  'Сделано объективом и любовью',
+  'Счастье никогда не выходит из моды',
+  'Лишь тот, кто странствует, открывает новые пути',
+  'Зарядитесь нашим теплом',
+  'Жизнь лучше, когда ты смеешься'];
+
+const COMMENT_TEXTS = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = ['Алиса','Владимир','Ибрагим','Дмитрий','Михаил','Аделина','Таисия','София','Тимофей','Егор','Кирилл','Лев','Валерия','Георгий','Андрей','Максим','Евгения','Матвей','Ксения','Анастасия','Ульяна','Дарья','Амина','Сабина','Артём','Мария','Варвара','Ян','Александра','Дарина','Елена','Иван','Степан','Евгений','Руслан','Арина','Алексей','Татьяна','Константин','Маргарита','Фатима','Анна','Алия','Ева','Виктория','Григорий','Герман','Елизавета','Ярослав','Роман'];
+
+const TEMPLATE = document.querySelector('#picture').content;
+
+export {PHOTO_NUMBER, MAX_COMMENT_NUMBER, PHOTO_DESCRIPTIONS, COMMENT_TEXTS, NAMES, TEMPLATE};
