+++ js/getRandomInt.js
@@ -1,0 +1,8 @@
+/* функция для случайного целого числа */
+function getRandomInt(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min) + min); // Максимум не включается, минимум включается
+}
+
+export {getRandomInt};
